refactors the FindHandler into two smaller classes
