+++ cg/store/api/core.py
@@ -4,6 +4,8 @@ import logging
 import alchy
 
 from cg.store import models
+from cg.store.api.find2 import FindHandler2
+
 from .add import AddHandler
 from .find import FindHandler
 from .status import StatusHandler
@@ -12,7 +14,7 @@ from .trends import TrendsHandler
 LOG = logging.getLogger(__name__)
 
 
-class CoreHandler(AddHandler, FindHandler, StatusHandler, TrendsHandler):
+class CoreHandler(AddHandler, FindHandler, FindHandler2, StatusHandler, TrendsHandler):
     """Aggregating class for the store api handlers"""
     pass
